backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
dspy-ai>=2.6.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
